+++ pkg/controller/reconciler/shared_test.go
@@ -18,13 +18,29 @@ package reconciler
 
 import (
 	"fmt"
+	"testing"
+	"time"
 
+	appsv1 "k8s.io/api/apps/v1"
 	v1 "github.com/radius-project/radius/pkg/armrpc/api/v1"
 	radappiov1alpha3 "github.com/radius-project/radius/pkg/controller/api/radapp.io/v1alpha3"
 	"github.com/radius-project/radius/pkg/corerp/api/v20231001preview"
 	"github.com/radius-project/radius/pkg/to"
+	"github.com/radius-project/radius/test/testcontext"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+const (
+	recipeTestWaitDuration            = time.Second * 10
+	recipeTestWaitInterval            = time.Second * 1
+	recipeTestControllerDelayInterval = time.Millisecond * 100
 )
 
 func createEnvironment(radius *mockRadiusClient, name string) {
@@ -48,4 +64,130 @@ func makeRecipe(name types.NamespacedName, resourceType string) *radappiov1alpha
 			Type: resourceType,
 		},
 	}
+}
+
+func waitForRecipeStateUpdating(t *testing.T, client client.Client, name types.NamespacedName, oldOperation *radappiov1alpha3.ResourceOperation) *radappiov1alpha3.RecipeStatus {
+	ctx := testcontext.New(t)
+
+	logger := t
+	status := &radappiov1alpha3.RecipeStatus{}
+	require.EventuallyWithT(t, func(t *assert.CollectT) {
+		logger.Logf("Fetching Recipe: %+v", name)
+		current := &radappiov1alpha3.Recipe{}
+		err := client.Get(ctx, name, current)
+		require.NoError(t, err)
+
+		status = &current.Status
+		logger.Logf("Recipe.Status: %+v", current.Status)
+		assert.Equal(t, status.ObservedGeneration, current.Generation, "Status is not updated")
+
+		if assert.Equal(t, radappiov1alpha3.PhraseUpdating, current.Status.Phrase) {
+			assert.NotEmpty(t, current.Status.Operation)
+			assert.NotEqual(t, oldOperation, current.Status.Operation)
+		}
+
+	}, recipeTestWaitDuration, recipeTestWaitInterval, "failed to enter updating state")
+
+	return status
+}
+
+func waitForRecipeStateReady(t *testing.T, client client.Client, name types.NamespacedName) *radappiov1alpha3.RecipeStatus {
+	ctx := testcontext.New(t)
+
+	logger := t
+	status := &radappiov1alpha3.RecipeStatus{}
+	require.EventuallyWithTf(t, func(t *assert.CollectT) {
+		logger.Logf("Fetching Recipe: %+v", name)
+		current := &radappiov1alpha3.Recipe{}
+		err := client.Get(ctx, name, current)
+		require.NoError(t, err)
+
+		status = &current.Status
+		logger.Logf("Recipe.Status: %+v", current.Status)
+		assert.Equal(t, status.ObservedGeneration, current.Generation, "Status is not updated")
+
+		if assert.Equal(t, radappiov1alpha3.PhraseReady, current.Status.Phrase) {
+			assert.Empty(t, current.Status.Operation)
+		}
+	}, recipeTestWaitDuration, recipeTestWaitInterval, "failed to enter updating state")
+
+	return status
+}
+
+func waitForRecipeStateDeleting(t *testing.T, client client.Client, name types.NamespacedName, oldOperation *radappiov1alpha3.ResourceOperation) *radappiov1alpha3.RecipeStatus {
+	ctx := testcontext.New(t)
+
+	logger := t
+	status := &radappiov1alpha3.RecipeStatus{}
+	require.EventuallyWithTf(t, func(t *assert.CollectT) {
+		logger.Logf("Fetching Recipe: %+v", name)
+		current := &radappiov1alpha3.Recipe{}
+		err := client.Get(ctx, name, current)
+		assert.NoError(t, err)
+
+		status = &current.Status
+		logger.Logf("Recipe.Status: %+v", current.Status)
+		assert.Equal(t, status.ObservedGeneration, current.Generation, "Status is not updated")
+
+		if assert.Equal(t, radappiov1alpha3.PhraseDeleting, current.Status.Phrase) {
+			assert.NotEmpty(t, current.Status.Operation)
+			assert.NotEqual(t, oldOperation, current.Status.Operation)
+		}
+	}, recipeTestWaitDuration, recipeTestWaitInterval, "failed to enter deleting state")
+
+	return status
+}
+
+func waitForRecipeDeleted(t *testing.T, client client.Client, name types.NamespacedName) {
+	ctx := testcontext.New(t)
+
+	logger := t
+	require.Eventuallyf(t, func() bool {
+		logger.Logf("Fetching Recipe: %+v", name)
+		current := &radappiov1alpha3.Recipe{}
+		err := client.Get(ctx, name, current)
+		if apierrors.IsNotFound(err) {
+			return true
+		}
+
+		logger.Logf("Recipe.Status: %+v", current.Status)
+		return false
+
+	}, recipeTestWaitDuration, recipeTestWaitInterval, "recipe still exists")
+}
+
+func makeDeployment(name types.NamespacedName) *appsv1.Deployment {
+	return &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        name.Name,
+			Namespace:   name.Namespace,
+			Annotations: map[string]string{},
+		},
+		Spec: appsv1.DeploymentSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"app": name.Name,
+				},
+			},
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: map[string]string{
+						"app": name.Name,
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name:  name.Name,
+							Image: "nginx:latest",
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func boolPtr(b bool) *bool {
+    return &b
 }
